refactor: Extract DirectApp and NuruApp into separate files
- Move DirectApp class from DirectAppBuilder.cs to DirectApp.cs
- Move NuruApp class from AppBuilder.cs to NuruApp.cs
- Follow single responsibility principle
- Makes codebase easier to navigate
