+++ Source/TimeWarp.Nuru/AppBuilder.cs
@@ -89,24 +89,3 @@ public class AppBuilder
     return new NuruApp(serviceProvider);
   }
 }
-
-/// <summary>
-/// A built TimeWarp.Console application.
-/// </summary>
-public class NuruApp
-{
-  private readonly IServiceProvider ServiceProvider;
-
-  public IServiceProvider Services => ServiceProvider;
-
-  public NuruApp(IServiceProvider serviceProvider)
-  {
-    ServiceProvider = serviceProvider;
-  }
-
-  public Task<int> RunAsync(string[] args)
-  {
-    NuruCli cli = ServiceProvider.GetRequiredService<NuruCli>();
-    return cli.RunAsync(args);
-  }
-}
